requirements.txt
dbt-snowflake==1.11.3
streamlit
snowflake-snowpark-python
pandas